User restrictions

+++ base/views.py
@@ -1,17 +1,51 @@
-import re
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
-
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 
 # Create your views here.
 
 # rooms = [
 #    {'id': 1, 'name': 'lets learn python.'}, 
-#    {'id': 2, 'name': 'Design with me'},
+#    {'id': 2, 'name': 'Design with me'},s
 #    {'id': 3, 'name': 'Frontend Developers'},
 # ]
+
+def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User doesnot exist')
+        
+        
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid credentials')
+            
+        
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('login')   
 
 def home(request):
     
@@ -36,6 +70,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -47,10 +82,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room!!') 
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -63,6 +101,9 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room!!')
+     
     if request.method == 'POST':
         room.delete()
         return redirect('home')
